Redesign, flash message, pagination working;

+++ database/migrations/2024_05_26_013611_remove_image_from_products_table.php
@@ -12,7 +12,9 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('products', function (Blueprint $table) {
-            $table->dropColumn('image');
+            if (Schema::hasColumn('products', 'image')) {
+                $table->dropColumn('image');
+            }
         });
     }
 
@@ -22,7 +24,9 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('products', function (Blueprint $table) {
-            $table->string('image')->nullable();
+            if (!Schema::hasColumn('products', 'image')) {
+                $table->string('image')->nullable();
+            }
         });
     }
 };
